+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Upgrade the `org.realityforge.com.google.jsinterop` artifact to version `1.0.0-b2-e6d791f`.
 * Upgrade the `org.realityforge.com.google.elemental2` artifacts to version `2.24`.
 * Remove `{@inheritDoc}` as it only explicitly indicates that the default behaviour at the expense of significant visual clutter.
+* Fix bug where `#` is not removed when routing to empty location.
 
 ### [v0.17](https://github.com/realityforge/router-fu/tree/v0.17) (2019-02-04)
 [Full Changelog](https://github.com/realityforge/router-fu/compare/v0.16...v0.17)